% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_dir}
\alias{cache_dir}
\title{Utilities for caching}
\usage{
cache_dir()
}
\value{
\code{cache_dir()} returns the path to the cache directory.
}
\description{
Some expensive operations are cached to avoid recomputing them.
Intermediate results are stored in a cache directory.
Users can set it via \code{habistats.cache_dir} option.
The default is set with \code{\link[tools:userdir]{tools::R_user_dir()}}.
}
\details{
General recommendation is to set it to a persistent directory with easy access,
such as \verb{~/.cache/habistats}, not a temporary directory like \code{\link[=tempdir]{tempdir()}}.
If you are working on a shared machine with other users,
setting the same cache directory will allow you to share the cache.
}
\examples{
old = options(habistats.cache_dir = "~/.cache/habistats-example")

habistats::cache_dir()

\dontrun{
  # Calculate the size
  fs::dir_info(habistats::cache_dir(), recurse = TRUE)$size |> sum()
}

options(old) # reset for this example, not needed in real use
}
