% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{index_brillouin}
\alias{index_brillouin}
\alias{count_values}
\title{Diversity Indices}
\usage{
index_brillouin(ras)

count_values(ras)
}
\arguments{
\item{ras}{A \code{stars} object.}
}
\description{
\code{\link[=index_brillouin]{index_brillouin()}} calculates the Brillouin diversity index for a raster layer.

\code{\link[=count_values]{count_values()}} counts the occurrence of values in a raster layer.
}
