% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn-geometry.R
\name{evaluate_iucn_range}
\alias{evaluate_iucn_range}
\title{Use geometry column of IUCN Spatial Data}
\usage{
evaluate_iucn_range(sci_name, iucn_source = NULL)
}
\arguments{
\item{sci_name}{A character vector of species names.}

\item{iucn_source}{Path to the directory containing IUCN shape files.
\code{iucn_source_path()} is used if \code{NULL} is given (default).}
}
\value{
\code{\link[=evaluate_iucn_range]{evaluate_iucn_range()}} returns a data.frame.
}
\description{
The geometry column of IUCN Spatial Data is evaluated in several ways.
Currently, perimeter, area, and some other features are calculated.
}
\examples{
\dontrun{
options(habistats.iucn_source = "path/to/your/data")
habistats::iucn_source_path()
}
old = options(habistats.cache_dir = "~/.cache/habistats-example")

sn = c("Strato arctica", "Diva satanica")
habistats::evaluate_iucn_range(sn)

options(old) # reset for this example, not needed in real use
}
