% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn-species.R
\name{iucn_spatial_species}
\alias{iucn_spatial_species}
\alias{iucn_species_gpkg}
\title{Species-level summary of IUCN Spatial Data}
\usage{
iucn_spatial_species(iucn_source = NULL)

iucn_species_gpkg(iucn_source = NULL, overwrite = FALSE)
}
\arguments{
\item{iucn_source}{Path to the directory containing IUCN shape files.
\code{iucn_source_path()} is used if \code{NULL} is given (default).}

\item{overwrite}{Whether to overwrite existing cache files.}
}
\value{
\code{iucn_spatial_species()} returns a data.frame extracted from \code{\link[=iucn_spatial_features]{iucn_spatial_features()}}.

\code{iucn_species_gpkg()} returns a data.frame with two columns,
\code{"sci_name"} and \code{"source"}, to be used for \code{\link[=evaluate_iucn_range]{evaluate_iucn_range()}}.
}
\description{
\code{\link[=iucn_species_gpkg]{iucn_species_gpkg()}} splits IUCN shape files into separate gpkg files
for each species, and returns their paths. This allows for parallel processing
in downstream analysis, and avoids the need to read large shape files again.

Use \code{\link[=iucn_spatial_features]{iucn_spatial_features()}} for quick skimming of the whole dataset,
and determine which species to analyze.
}
