% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn.R
\name{iucn_spatial_features}
\alias{iucn_spatial_features}
\alias{iucn_spatial_species}
\title{Read and process IUCN Spatial Data}
\usage{
iucn_spatial_features(iucn_source = NULL)

iucn_spatial_species(iucn_source = NULL)
}
\arguments{
\item{iucn_source}{Path to the directory containing IUCN shape files.
\code{getOption("habistats.iucn_source")} is used if \code{NULL} is given (default).}
}
\description{
The IUCN Spatial Data is a large collection of shape files, and these functions
provide an efficient way to read and process them.
It may take a while to run them for the first time,
but subsequent runs will be very quick because the results are cached.

\code{\link[=iucn_spatial_features]{iucn_spatial_features()}} returns a data.frame without the geometry column.
This is useful for quick skimming of the whole dataset.

\code{\link[=iucn_spatial_species]{iucn_spatial_species()}} splits IUCN shape files into separate gpkg files
for each species, and returns their paths. This allows for parallel processing
in downstream analysis, and avoids the need to read large shape files again.
}
\seealso{
\url{https://www.iucnredlist.org/resources/spatial-data-download}
}
