% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koeppen.R
\name{raster_kmz}
\alias{raster_kmz}
\alias{raster_kgc}
\title{Köppen-Geiger Climate Classification}
\usage{
raster_kmz(mask = NULL)

raster_kgc(mask = NULL)
}
\arguments{
\item{mask}{sfg object.}
}
\description{
\code{\link[=raster_kmz]{raster_kmz()}} returns the KGC raster masked by a given polygon.
}
\seealso{
\url{https://koeppen-geiger.vu-wien.ac.at}
}
