% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{rowwise_map}
\alias{rowwise_map}
\alias{rowwise_map_lgl}
\alias{rowwise_map_int}
\alias{rowwise_map_dbl}
\alias{rowwise_map_chr}
\alias{rowwise_map_vec}
\alias{rowwise_mcmap}
\alias{rowwise_mcmap_lgl}
\alias{rowwise_mcmap_int}
\alias{rowwise_mcmap_dbl}
\alias{rowwise_mcmap_chr}
\alias{rowwise_mcmap_vec}
\title{Rowwise operations}
\usage{
rowwise_map(
  .x,
  .f,
  mode = c("list", "integer", "double", "character", "logical")
)

rowwise_map_lgl(.x, .f)

rowwise_map_int(.x, .f)

rowwise_map_dbl(.x, .f)

rowwise_map_chr(.x, .f)

rowwise_map_vec(.x, .f)

rowwise_mcmap(.x, .f, ...)

rowwise_mcmap_lgl(.x, .f, ...)

rowwise_mcmap_int(.x, .f, ...)

rowwise_mcmap_dbl(.x, .f, ...)

rowwise_mcmap_chr(.x, .f, ...)

rowwise_mcmap_vec(.x, .f, ...)
}
\arguments{
\item{.x}{A data frame-like object.}

\item{.f}{A function to apply to each row.}

\item{mode}{Output type.}

\item{...}{Additional arguments passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}.}
}
\description{
\verb{rowwise_*()} functions apply a function to each row of a data.frame.
The input to the function \code{.f} is a single-row subset with the same class.
This is useful for handling subclasses of data.frame, such as \code{sf}.
\code{base::apply(.x, 1L, .f)} cannot do this because it converts the input to a matrix.

\verb{rowwise_mcmap*()} are parallelized versions of \verb{rowwise_map*()}.
}
\seealso{
\code{\link[=lmap_lst]{lmap_lst()}}
}
