% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{summarize_raster}
\alias{summarize_raster}
\alias{count_values}
\title{Diversity Indices}
\usage{
summarize_raster(ras)

count_values(ras)
}
\arguments{
\item{ras}{A \code{stars} object.}
}
\description{
\code{\link[=summarize_raster]{summarize_raster()}} calculates various diversity indices from a raster layer.

\code{\link[=count_values]{count_values()}} counts the occurrence of values in a raster layer.
}
