% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habistats-package.R
\docType{package}
\name{habistats-package}
\alias{habistats}
\alias{habistats-package}
\title{habistats: Habitat Statistics}
\description{
Summarize and visualize the statistics of spatial data. Ecology and evolutionary biology are the main targets of this package, and utilites to handle IUCN Red List Spatial Data and Köppen-Geiger Climate Classification are provided. But the core part of the package is data-agnostic, and can be used with any spatial data, even if they are not related to habitat or biology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/habistats}
  \item \url{https://heavywatal.github.io/habistats/}
  \item Report bugs at \url{https://github.com/heavywatal/habistats/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

Authors:
\itemize{
  \item Fuku Sakamoto \email{fukusakamoto29@gmail.com} (\href{https://orcid.org/0009-0009-8531-5303}{ORCID})
}

}
\keyword{internal}
