% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habistats-package.R
\docType{package}
\name{habistats-package}
\alias{habistats}
\alias{habistats-package}
\title{habistats: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/habistats}
  \item \url{https://heavywatal.github.io/habistats/}
  \item Report bugs at \url{https://github.com/heavywatal/habistats/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
