% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn.R
\name{iucn_spatial_features}
\alias{iucn_spatial_features}
\alias{iucn_source_path}
\title{Read and process IUCN Spatial Data}
\usage{
iucn_spatial_features(iucn_source = NULL, overwrite = FALSE)

iucn_source_path()
}
\arguments{
\item{iucn_source}{Path to the directory containing IUCN shape files.
\code{iucn_source_path()} is used if \code{NULL} is given (default).}

\item{overwrite}{Whether to overwrite existing cache files.}
}
\value{
\code{iucn_spatial_features()} returns a data.frame with the feature columns.

\code{iucn_source_path()} returns a path read from the option.
}
\description{
The IUCN Spatial Data is a large collection of shape files,
and this package provides efficient ways to read and process them.
It may take a while to run them for the first time,
but subsequent runs will be very quick because the results are cached.

\code{iucn_spatial_features()} reads the IUCN shape files and drops the geometry column.
This is useful for quick skimming of the whole dataset.
Set \code{options(habistats.iucn_source = "~/path/to/your/data")} to specify
the path to the directory containing IUCN shape files,
and use \code{iucn_source_path()} to check it.
}
\seealso{
\url{https://www.iucnredlist.org/resources/spatial-data-download}
}
