% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koeppen.R
\name{raster_kgc}
\alias{raster_kgc}
\title{Köppen-Geiger Climate Classification}
\format{
A \code{stars} object with 2 dimensions and 1 attribute.
}
\source{
\url{https://koeppen-geiger.vu-wien.ac.at}
}
\usage{
raster_kgc(mask = NULL)
}
\arguments{
\item{mask}{sf or sfc object.}
}
\description{
\code{\link[=raster_kgc]{raster_kgc()}} returns the KGC raster masked by a given polygon.
}
