% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koeppen.R
\name{raster_kgc}
\alias{raster_kgc}
\alias{scale_color_kgc}
\alias{scale_fill_kgc}
\title{Köppen-Geiger Climate Classification}
\format{
A \code{stars} object with 2 dimensions and 1 attribute.
}
\source{
\url{https://koeppen-geiger.vu-wien.ac.at}
}
\usage{
raster_kgc(mask = NULL)

scale_color_kgc(...)

scale_fill_kgc(...)
}
\arguments{
\item{mask}{sf or sfc object.}

\item{...}{passed to \code{\link[ggplot2:scale_manual]{ggplot2::scale_fill_manual()}}.}
}
\value{
\code{raster_kgc()} returns a \code{stars} object.

\code{scale_color_kgc()} and \code{scale_fill_kgc()} return a ggplot2 scale.
}
\description{
\code{raster_kgc()} gives easy access to the Köppen-Geiger climate classification
raster data. It can be cropped by a given polygon.
\code{scale_color_kgc()} and \code{scale_fill_kgc()} are ggplot2 color scales
to be used with the KGC raster.
}
\examples{
library(ggplot2)
geo_jp = rnaturalearth::ne_countries(country = "Japan")
kgc_jp = habistats::raster_kgc(geo_jp) |> print()
ggplot(geo_jp) +
  stars::geom_stars(data = kgc_jp) +
  geom_sf(linewidth = 1, color = "#00000088", fill = NA) +
  scale_fill_kgc(na.value = "transparent") +
  theme_minimal()
}
