% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-raster.R
\name{quantify_raster}
\alias{quantify_raster}
\alias{get_values_na_omit}
\title{Qantify spatial data in raster format}
\usage{
quantify_raster(ras)

get_values_na_omit(ras)
}
\arguments{
\item{ras}{A \code{stars} object.}
}
\value{
\code{quantify_raster()} returns a data.frame.

\code{count_values()} returns an integer vector.
}
\description{
\code{quantify_raster()} calculates various diversity indices from a raster layer.
}
