% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify-vector.R
\name{quantify_vector_kgc}
\alias{quantify_vector_kgc}
\alias{quantify_vector}
\title{Quantify spatial data in vector format}
\usage{
quantify_vector_kgc(geom)

quantify_vector(geom)
}
\arguments{
\item{geom}{An \code{sfc} or \code{sf} object. A path to sf-readable file is accepted.}
}
\value{
A data.frame.
}
\description{
\code{quantify_vector()} calculates the basic spatial features of a vector geometry
such as perimeter and area.
\code{quantify_vector_kgc()}, in addition, quantifies the Köppen-Geiger climate
classification in that range cropped with \code{\link[=raster_kgc]{raster_kgc()}}.
}
\examples{
geom = rnaturalearth::ne_countries(country = "Japan") |> sf::st_geometry()
quantify_vector_kgc(geom)
}
