% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{rowwise_map}
\alias{rowwise_map}
\alias{rowwise_map_lgl}
\alias{rowwise_map_int}
\alias{rowwise_map_dbl}
\alias{rowwise_map_chr}
\alias{rowwise_map_vec}
\alias{rowwise_mcmap}
\alias{rowwise_mcmap_lgl}
\alias{rowwise_mcmap_int}
\alias{rowwise_mcmap_dbl}
\alias{rowwise_mcmap_chr}
\alias{rowwise_mcmap_vec}
\title{Rowwise operations}
\usage{
rowwise_map(
  .x,
  .f,
  mode = c("list", "integer", "double", "character", "logical")
)

rowwise_map_lgl(.x, .f)

rowwise_map_int(.x, .f)

rowwise_map_dbl(.x, .f)

rowwise_map_chr(.x, .f)

rowwise_map_vec(.x, .f, .ptype = NULL)

rowwise_mcmap(.x, .f, ...)

rowwise_mcmap_lgl(.x, .f, ...)

rowwise_mcmap_int(.x, .f, ...)

rowwise_mcmap_dbl(.x, .f, ...)

rowwise_mcmap_chr(.x, .f, ...)

rowwise_mcmap_vec(.x, .f, ..., .ptype = NULL)
}
\arguments{
\item{.x}{A data frame-like object.}

\item{.f}{A function to apply to each row.}

\item{mode}{Output type.}

\item{.ptype}{A prototype for the output vector.}

\item{...}{Additional arguments passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}.}
}
\value{
A vector of the same length as the number of rows in \code{.x}.
The suffix of the function name denotes the output type.
\verb{*_vec()} applies \code{\link[purrr:list_simplify]{purrr::list_simplify()}} before returning.
}
\description{
\verb{rowwise_*()} functions apply a function to each row of a data.frame.
The input to the function \code{.f} is a single-row subset with the same class.
This is useful for handling subclasses of data.frame, such as \code{sf}.
\code{base::apply(.x, 1L, .f)} cannot do this because it converts the input to a matrix.

\verb{rowwise_mcmap*()} are parallelized versions of \verb{rowwise_map*()}.
}
\examples{
df = data.frame(a = 1:2, b = 3:4)
rowwise_map(df, identity)

rowwise_map_lgl(df, is.data.frame)

rowwise_map_int(df, ncol)

rowwise_map_dbl(df, sum)

rowwise_map_chr(df, toString)

rowwise_map_vec(df, \(x) x$a + x$b * 1i)
}
\seealso{
\code{\link[=lmap_lst]{lmap_lst()}}
}
